feat: add @hi-ui/classname util
