+++ .gitignore
@@ -1,8 +1,7 @@
 # MIFE dotfiles
 
 # Output files
-packages/*/lib/
-packages/*/es/
+lib/
 
 # Editor generate files
 .idea/
