+++ packages/utils/classname/README.md
@@ -1,0 +1,11 @@
+# `@hi-ui/classname`
+
+> TODO: description
+
+## Usage
+
+```
+const classname = require('@hi-ui/classname');
+
+// TODO: DEMONSTRATE API
+```
