+++ packages/utils/classname/__tests__/classname.test.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const classname = require('..')
+
+describe('@hi-ui/classname', () => {
+  it('needs tests')
+})
