+++ packages/utils/classname/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@hi-ui/classname",
+  "version": "1.0.0",
+  "description": "Now I’m the model of a modern major general / The venerated Virginian veteran whose men are all / Lining up, to put me up on a pedestal / Writin’ letters to relatives / Embellishin’ my elegance and eloquence / But the elephant is in the room / The truth is in ya face when ya hear the British cannons go / BOOM",
+  "keywords": [],
+  "author": "Flcwl <flcwl@vip.qq.com>",
+  "license": "MIT",
+  "main": "lib/classname.js",
+  "directories": {
+    "lib": "lib",
+    "test": "__tests__"
+  },
+  "files": [
+    "lib"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/XiaoMi/hiui.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1",
+    "prebuild": "rimraf lib",
+    "build:esm": "cross-env BABEL_MODULE_ENV=esm hi-build ./src/index.ts --format esm -d ./lib/esm",
+    "build:cjs": "cross-env BABEL_MODULE_ENV=cjs hi-build ./src/index.ts --format cjs -d ./lib/cjs",
+    "build:types": "tsc --emitDeclarationOnly --declarationDir lib/types",
+    "build": "concurrently yarn:build:*"
+  },
+  "bugs": {
+    "url": "https://github.com/XiaoMi/hiui/issues"
+  },
+  "homepage": "https://github.com/XiaoMi/hiui#readme",
+  "dependencies": {
+    "classnames": "^2.3.1"
+  }
+}
